Move mongo storage implementation so we can make another
